init: original creation of flux compspec module
Problem: we want to use the rpc to create a node list.
Solution: create a compspec Python plugin. Note that this
is potentially deviant in behavior because I was not able
to use the RPC to generate a full graph, but just the
shorthand. I tried to account for all cases - either
finding the shorthand and expanding it, finding
the v1 graph and using it, either converting to v2 JGF
or returning as is under direction of the user. I do
not have need for this right now but wanted to create
it in case future folks do.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>