feat: add testing and standalone mode
Problem: we should have a basic test that the plugin works
with compspec.
Solution: add a test, along with refactoring the plugin
organization so that it is possible to run a single
function in a flux instance, "generate_cluster" that will
dump out the node graph on the fly, and no need to be
executed as an extraction plugin by the higher level
compspec.

Signed-off-by: vsoch <vsoch@users.noreply.github.com>